Реализация класса первой работы
